move test to private
